Fix NVidia Threaded Optimizations on Linux
This makes Nvidia Driver behave as it does in Windows when setting the env var
__GL_THREADED_OPTIMIZATIONS=1

It makes OpenGL Calls work in a non blocking way when and be executed in
a separate thread than the MTGS.

- Brought back GSWndOGL (GLX) implementation since the optimizations don't
seem to work with EGL.

- Call XInitThreads() before any other thing, see
https://download.nvidia.com/XFree86/Linux-x86_64/460.91.03/README/openglenvvariables.html:

"Additionally, these optimizations require Xlib to function in
thread-safe mode. The NVIDIA OpenGL driver cannot reliably enable Xlib
thread-safe mode itself, therefore the application needs to call
XInitThreads() before making any other Xlib call. Otherwise, the
threaded optimizations in the NVIDIA driver will not be enabled."